requirements.txt
python-dotenv==1.0.0
langchain==0.1.0
openai
google-api-python-client
google-search-results
youtube-transcript-api
chromadb
tiktoken
faiss-cpu
wikipedia
psycopg2-binary
pinecone-client
pgvector
Flask==2.3.4
flask-cors==4.0.0
langchain-openai==0.1.0
